backend/requirements.txt
asgiref==3.8.1
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.1
cycler==0.12.1
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
fonttools==4.55.3
google-auth==2.36.0
google-auth-oauthlib==1.2.1
gspread==6.1.4
gunicorn==23.0.0
idna==3.10
kiwisolver==1.4.7
matplotlib==3.10.0
numpy==2.1.3
oauthlib==3.2.2
packaging==24.2
pandas==2.2.3
pillow==11.0.0
psycopg2-binary==2.9.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
sqlparse==0.5.2
tzdata==2024.2
urllib3==2.2.3
